Safari-609.1.11.1/LayoutTests/fast/history/page-cache-webdatabase-pending-transaction-expected.txt

CONSOLE MESSAGE: line 54: Web SQL is deprecated. Please use IndexedDB instead.
Tests that a page with an open WebDatabase that has pending transactions is able to go into the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the back/forward cache
PASS All done
PASS successfullyParsed is true

TEST COMPLETE
